Removed LinearRing and Line
LineString does not have subclasses any more.
One can check if it forms a linear ring by calling `isRing()`, and if it forms a line by checking `numPoints() == 2`.

This introduces the ability to form potentially invalid polygons with self-intersecting rings, and leaves the developer the choice to explicitly check them for validity with the new `isValid()` method introduced earlier. This is the approach followed by most implementations, including database vendors and the GEOS project.

This consequently speeds up creation of LineString and therefore Polygon objects, by not systematically running these expensive computations on instantiation.